Acitelli

Acitelli, L. K. (1997). Sampling couples to understand them: Mixing the theoretical with the practical. Journal of Social and Personal Relationships, 14, 243-261.

acitelli

Format

A data frame with 296 rows and 16 variables:

CoupleID

each couple has 1 couple id

Yearsmar

each couple has 1 value for years married (centered; 11.214 subtracted)

Partnum

A variable to distinguish the two members (1 and 2)

Gender_A

-1 Female; 1 Male

SelfPos_A

Mean of positive self variables (uncentered)

OtherPos_A

Mean of positive partner variables (uncentered)(both variables the average of Cooperative, Mature, Friendly, Hard Working and Care about Others)

Satisfaction_A

Mean of four variables (theoretical range from 1 to 4)

Tension_A

Mean of two disagreement and tension (theoretical range from 1 to 4)

SimHob_A

Perception of having similar hobbies (1 = very true; 0 = somewhat true; -1 = not at all true)

Gender_P

-1 Female; 1 Male

SelfPos_P

Mean of positive self variables (uncentered)

OtherPos_P

Mean of positive partner variables (uncentered) (both variables the average of Cooperative, Mature, Friendly, Hard Working and Care about Others)

Satisfaction_P

Mean of four variables (theoretical range from 1 to 4)

Tension_P

Mean of two disagreement and tension (theoretical range from 1 to 4)

SimHob_P

Perception of having similar hobbies (1 = very true; 0 = somewhat true; -1 = not at all true)

Gender

String variable with Husband and Wife

Source

https://journals.sagepub.com/doi/10.1177/0265407597142006