Confirmatory Factor Analysis (groupwise)

This function will run a series of CFA (n = length(group)) with respect to each group. The function is intended to help you get a better understanding of which group has abnormal fit indicator

cfa_groupwise(data, ..., model = NULL, group, ordered = F)

Arguments

data
data frame

...
CFA items.

model
explicit lavaan model. Must be specify with model = lavaan_model_syntax.

group
character. group variable.

ordered
logical. default is F. If it is set to T, lavaan will treat it as a ordinal variable and use DWLS instead of ML

Details

All argument must be explicitly specified. If not, all arguments will be treated as CFA items

References

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel

Rosseel Y (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1–36. https://www.jstatsoft.org/v48/i02/.

Examples

# The example is used as the illustration of the function output only.
# It does not imply the data is appropriate for the analysis.
cfa_groupwise(
  data = lavaan::HolzingerSwineford1939,
  group = "school",
  x1:x3,
  x4:x6,
  x7:x9
)

#>         group       cfi      rmsea       tli
#> 1     Pasteur 0.9031037 0.10376088 0.8546555
#> 2 Grant-White 0.9413704 0.08896312 0.9120556