Confirmatory Factor Analysis

The function fits a CFA model using the lavaan::cfa (Rosseel, 2012) function. Users can fit single and multiple factors CFA (read details), and it also supports multilevel CFA (specifying the group). Users can pass the items (see below for example) or an explicit lavaan model for more versatile usage.

cfa_summary(
  data,
  ...,
  model = NULL,
  group = NULL,
  summary_item = c("cfi", "rmsea", "tli", "srmr"),
  ordered = F,
  return_result = "model",
  quite = F,
  group_partial = NULL
)

Arguments

data
data frame

...
CFA items. Multi-factor CFA items should be separated by comma (as different argument). See below for example

model
explicit lavaan model. Must be specify with model = lavaan_model_syntax.

group
optional character. used for multi-level CFA. the nested variable for multilevel dataset (e.g., Country)

summary_item
vector of fit indices. Default is CFI, RMSEA, TLI, and SRMR.

ordered
logical. default is F. If it is set to T, lavaan will treat it as a ordinal variable and use DWLS instead of ML

return_result
Default is model. Options are 'model' (lavaan model), 'short_summary' (fit index summary only), 'long_summary' (lavaan full summary).

quite
default as F. If set to true, it will not print the running model statement.

group_partial
Items for partial equivalence. The form should be c('DV =~ item1', 'DV =~ item2').

Details

All argument must be explicitly specified. If not, all arguments will be treated as CFA items.

References

Bao, H.-W.-S. (2021). bruceR: Broadly useful convenient and efficient R functions. R package version 0.6.0. https://CRAN.R-project.org/package=bruceR

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel

Rosseel Y (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1–36. https://www.jstatsoft.org/v48/i02/.

Examples

# REMEMBER, YOU MUST NAMED ALL ARGUMENT EXCEPT THE CFA ITEMS ARGUMENT
# Fitting a single factor CFA model
cfa_summary(
  data = lavaan::HolzingerSwineford1939,
  x1:x3
)

#> Computing CFA using:
#>   DV1 =~ x1 + x2 + x3
#>  

#> lavaan 0.6-8 ended normally after 23 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                         6
#>                                                       
#>   Number of observations                           301
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0

# Fitting a multilevel single factor CFA model
cfa_summary(
  data = lavaan::HolzingerSwineford1939,
  x1:x3,
  group = "sex"
)

#> Computing CFA using:
#>   DV1 =~ x1 + x2 + x3
#>  

#> lavaan 0.6-8 ended normally after 45 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        18
#>                                                       
#>   Number of observations per group:                   
#>     1                                              146
#>     2                                              155
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                 0.000
#>   Degrees of freedom                                 0
#>   Test statistic for each group:
#>     1                                            0.000
#>     2                                            0.000

# Fitting a multiple factor CFA model
cfa_summary(
  data = lavaan::HolzingerSwineford1939,
  x1:x3,
  x4:x6,
  x7:x9
)

#> Computing CFA using:
#>   DV1 =~ x1 + x2 + x3
#>   DV2 =~ x4 + x5 + x6
#>   DV3 =~ x7 + x8 + x9
#>  

#> lavaan 0.6-8 ended normally after 35 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        21
#>                                                       
#>   Number of observations                           301
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                85.306
#>   Degrees of freedom                                24
#>   P-value (Chi-square)                           0.000

# Fitting a CFA model by passing explicit lavaan model (equivalent to the above model)
# Note in the below function how I added `model = ` in front of the lavaan model.
# Similarly, the same rule apply for all arguments (e.g., `ordered = F` instead of `F`)
cfa_summary(
  model = "visual  =~ x1 + x2 + x3
                     textual =~ x4 + x5 + x6
                     speed   =~ x7 + x8 + x9 ",
  data = lavaan::HolzingerSwineford1939
)

#> Computing CFA using:
#>  visual  =~ x1 + x2 + x3
#>                      textual =~ x4 + x5 + x6
#>                      speed   =~ x7 + x8 + x9 

#> lavaan 0.6-8 ended normally after 35 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        21
#>                                                       
#>   Number of observations                           301
#>                                                       
#> Model Test User Model:
#>                                                       
#>   Test statistic                                85.306
#>   Degrees of freedom                                24
#>   P-value (Chi-square)                           0.000

if (FALSE) {
# This will fail because I did not add `model = ` in front of the lavaan model.
# Therefore,you must add the tag in front of all arguments
# For example, `return_result = 'model'` instaed of `model`
cfa_summary("visual  =~ x1 + x2 + x3
             textual =~ x4 + x5 + x6
             speed   =~ x7 + x8 + x9 ",
  data = lavaan::HolzingerSwineford1939
)
}