Mean, SD, and Correlation table

This function generates a table of descriptive statistics (mainly using psych::describe; Revelle, 2021) and or a correlation table. User can export this to a csv file (optionally, using the file_path argument). Users can open the csv file with MS Excel then copy and paste the table into MS Word table.

descriptive_table(
  data,
  cols,
  cor_sig_test = "raw",
  cor_digit = 3,
  descriptive_indicator = c("mean", "sd", "cor"),
  descriptive_indicator_digit = 3,
  file_path = NULL
)

Arguments

data
data frame

cols
vector or tidyselect syntax or helpers. column(s) need to be included in the table.

cor_sig_test
adjusted or raw. Default as adjusted. See psych::corr.test to learn more.

cor_digit
number of digit for correlation table

descriptive_indicator
Default is mean, sd, cor. Options are missing (missing value count), non_missing (non-missing value count), cor (correlation table), n, mean, sd, median, trimmed (trimmed mean), median, mad (median absolute deviation from the median), min, max, range, skew, kurtosis, se (standard error)

descriptive_indicator_digit
number of digit for the descriptive table

file_path
file path for export. The function will implicitly pass this argument to the write.csv(file = file_path)

References

Revelle, W. (2021). psych: Procedures for Psychological, Psychometric, and Personality Research. Northwestern University, Evanston, Illinois. R package version 2.1.3, https://CRAN.R-project.org/package=psych.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel

Examples

descriptive_table(iris, cols = where(is.numeric)) # all numeric columns

#>               mean    sd Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length 5.843 0.828                                                  
#> Sepal.Width  3.057 0.436       -0.118                                     
#> Petal.Length 3.758 1.765     0.872***   -0.428***                         
#> Petal.Width  1.199 0.762     0.818***   -0.366***     0.963***            

descriptive_table(iris,
  cols = where(is.numeric),
  # get missing count, non-missing count, and mean & sd & correlation table
  descriptive_indicator = c("missing", "non_missing", "mean", "sd", "cor")
)

#>              missing_n non_missing_n  mean    sd Sepal.Length Sepal.Width
#> Sepal.Length         0           150 5.843 0.828                         
#> Sepal.Width          0           150 3.057 0.436       -0.118            
#> Petal.Length         0           150 3.758 1.765     0.872***   -0.428***
#> Petal.Width          0           150 1.199 0.762     0.818***   -0.366***
#>              Petal.Length Petal.Width
#> Sepal.Length                         
#> Sepal.Width                          
#> Petal.Length                         
#> Petal.Width      0.963***