Generalized Linear Mixed Effect Model

Generalized linear mixed effect model. The function uses the lme4::glmer(Bates et al., 2014) function. This function is still in very early development stage.

glme_model(
  data,
  response_variable,
  level_1_factors,
  level_2_factors = NULL,
  family,
  two_way_interaction_factor = NULL,
  three_way_interaction_factor = NULL,
  id,
  estimation_method = "REML",
  opt_control = "bobyqa",
  na.action = stats::na.exclude,
  quite = F
)

Arguments

data
data frame

response_variable
character or vector of length 1

level_1_factors
vector. Level-1 variables (e.g., individual-level)

level_2_factors
optional vector. level-2 variables (e.g., group-level)

family
a GLM family. It will passed to the family argument in glmer. See ?glmer for possible options.

two_way_interaction_factor
optional vector of length more than 2. Default to null

three_way_interaction_factor
optional vector of length 3. Do not include two-way interaction factors if this is not null. Default to null

id
character or vector of length 1. The nesting variable (e.g. group)

estimation_method
character. ML or REML default to REML.

opt_control
character. default to bobyqa

na.action
default to stats::na.exclude.

quite
default to F. If set to T, it will not print the fitting model statement

Value

An object of class "glmerMod" representing the linear mixed-effects model fit.

References

Bates, D., Mächler, M., Bolker, B., & Walker, S. (2014). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1–48. doi: 10.18637/jss.v067.i01.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel

Examples

x <- rnorm(200)
y <- rpois(200, exp(1 + x))
group <- rep(1:10, 20)
test_df <- as.data.frame(cbind(x, y, group))
glme_model(
  data = test_df,
  response_variable = "y",
  level_1_factors = "x",
  family = poisson(link = "log"),
  id = "group"
)

#> Fitting Model with glmer: 
#>  Formula = y ~ x + (1 + x | group)
#>  Family = poisson

#> boundary (singular) fit: see ?isSingular

#> Generalized linear mixed model fit by maximum likelihood (Laplace
#>   Approximation) [glmerMod]
#>  Family: poisson  ( log )
#> Formula: y ~ x + (1 + x | group)
#>    Data: 
#> structure(list(x = c(-1.40004351672175, 0.25531705484526, -2.43726361121953,  
#> -0.00557128674616073, 0.621552721415214, 1.14841160602606, -1.82181766097663,  
#> -0.247325302073524, -0.244199606778383, -0.282705448814465, -0.553699383688721,  
#> 0.628982042036008, 2.06502489535922, -1.63098940208223, 0.512426949851805,  
#> -1.86301149206833, -0.522012514745454, -0.0526019099538795, 0.54299634266114,  
#> -0.914074827259928, 0.468154420450533, 0.362951255864986, -1.30454354503478,  
#> 0.737776321255047, 1.88850492923455, -0.0974451044082059, -0.935847353500678,  
#> -0.0159503112505377, -0.826788953733443, -1.5123996512628, 0.935363189801085,  
#> 0.176488610562459, 0.243685464746205, 1.6235488834436, 0.112038082869763,  
#> -0.133997013062555, -1.91008746772221, -0.279237241584845, -0.31344597825789,  
#> 1.06730787876151, 0.0700348502966383, -0.639123323885664, -0.0499648990481453,  
#> -0.251483443250214, 0.444797115792728, 2.75541757533686, 0.0465313804441933,  
#> 0.577709069306631, 0.118194874409683, -1.9117204910282, 0.862086481838568,  
#> -0.243236739633475, -0.206087194781581, 0.0191775922091463, 0.0295607543197664,  
#> 0.549827541801689, -2.27411485689294, 2.68255718355635, -0.361221255493715,  
#> 0.21335574971003, 1.07434588225969, -0.665088248545505, 1.11395241896922,  
#> -0.245896411679153, -1.17756330894328, -0.975850616159863, 1.06505732002715,  
#> 0.131670634772894, 0.488628809200981, -1.69945056784569, -1.47073630637743,  
#> 0.284150343626922, 1.33732041323558, 0.236696283286503, 1.31829338403463,  
#> 0.523909787501208, 0.606748046557118, -0.10993567214487, 0.172181715411,  
#> -0.0903272870006624, 1.92434334135779, 1.29839275883088, 0.748791267536843,  
#> 0.556224329104296, -0.548257263738865, 1.11053489258185, -2.6123343328843,  
#> -0.155693776230052, 0.433889789917249, -0.381951111816166, 0.424187574547461,  
#> 1.06310199584878, 1.04871261975573, -0.0381028948131284, 0.486148919515245,  
#> 1.67288261083248, -0.354361163556443, 0.94634788646655, 1.31682635604417,  
#> -0.296640024591827, -0.387213575248403, -0.785432655780338, -1.05673686699317,  
#> -0.795541429547809, -1.75627542817088, -0.690537896821751, -0.558541994218778,  
#> -0.536663326497313, 0.227127132728239, 0.97845492024324, -0.208882651351827,  
#> -1.39941046001177, 0.258537287908512, -0.441799452911353, 0.56859986121534,  
#> 2.12685045903409, 0.424858441314262, -1.68428153224971, 0.249401783969334,  
#> 1.07283825181642, 2.03936926250284, 0.44945377806039, 1.39181404563834,  
#> 0.426566546903553, 0.107583992213665, 0.0222947331283422, 0.603611010507109,  
#> -0.262650572849392, -0.528264082370987, 0.192149422037578, -1.14619966899296,  
#> 0.846184664797661, 0.0817196292515542, -1.30511700993532, -0.944912059949758,  
#> 0.454341594315134, -0.85520250080836, -0.286895218789768, 0.894961625519496,  
#> 0.067304439692084, -0.162676337206272, -0.827310169311162, 1.87650562102549,  
#> 0.766440198879035, 0.979956695586349, 1.32178099162845, -1.11971082861119,  
#> 0.514599818539648, -1.50909983576966, 1.53274148032248, 0.429147371132052,  
#> 0.122103414472254, -1.13801240059386, -0.558015129418299, 1.05253853690161,  
#> 0.677683644301391, 0.0384995470644212, -0.356381187004009, 0.782844101839823,  
#> 0.804411616129238, -1.90006082335695, 0.93578428615882, -0.309051503430768,  
#> 0.263066676962616, -1.79059185585806, -0.788258844612912, -1.13302166872061,  
#> 0.363652568373227, -0.285887914303066, 0.517669133997423, -0.102908670119104,  
#> -0.974069593265456, 1.27067230079593, 0.960864787077774, 0.768721369594144,  
#> 1.03593077108096, -0.473887074273069, -1.27533487463628, -0.305620673829671,  
#> 2.21176948683873, -1.0416683808784, -1.14652385021424, -1.67532730336442,  
#> 1.52593865503731, 0.554185514898162, 1.99311026486264, -0.154120739556124,  
#> 2.56440833799741, 1.06199914454589, 1.14269487842917, 1.12383884326229,  
#> -0.397001493059625, -0.823261150567127, -0.5788846247121, 1.7637893775715,  
#> 0.132992146404077, 0.376499328478288, 1.13870765340302, 1.24126307530705,  
#> 0.612090944959137), y = c(0, 5, 1, 1, 5, 8, 0, 1, 0, 1, 1, 7,  
#> 16, 0, 1, 0, 6, 5, 2, 0, 3, 4, 1, 5, 12, 2, 1, 2, 2, 2, 8, 5,  
#> 1, 18, 4, 2, 0, 2, 0, 8, 3, 1, 3, 1, 3, 50, 3, 3, 2, 0, 8, 0,  
#> 0, 2, 3, 3, 0, 44, 1, 1, 10, 1, 6, 1, 1, 0, 9, 4, 3, 1, 1, 1,  
#> 9, 6, 14, 7, 1, 4, 3, 0, 20, 13, 6, 5, 1, 12, 0, 2, 4, 1, 2,  
#> 4, 6, 7, 2, 16, 0, 11, 14, 1, 2, 1, 0, 0, 0, 2, 4, 0, 4, 10,  
#> 0, 0, 3, 1, 4, 22, 1, 0, 5, 14, 21, 1, 10, 4, 4, 3, 5, 0, 4,  
#> 3, 3, 6, 2, 1, 0, 9, 1, 6, 7, 4, 2, 2, 19, 5, 10, 9, 3, 6, 0,  
#> 13, 5, 4, 2, 1, 5, 3, 3, 3, 10, 7, 0, 2, 2, 5, 0, 1, 0, 1, 3,  
#> 4, 5, 2, 5, 9, 3, 5, 3, 1, 3, 22, 0, 1, 0, 14, 7, 25, 2, 42,  
#> 10, 10, 11, 5, 0, 4, 17, 3, 3, 9, 6, 6), group = c(1, 2, 3, 4,  
#> 5, 6, 7, 8, 9, 10, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 1, 2, 3, 4,  
#> 5, 6, 7, 8, 9, 10, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 1, 2, 3, 4,  
#> 5, 6, 7, 8, 9, 10, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 1, 2, 3, 4,  
#> 5, 6, 7, 8, 9, 10, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 1, 2, 3, 4,  
#> 5, 6, 7, 8, 9, 10, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 1, 2, 3, 4,  
#> 5, 6, 7, 8, 9, 10, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 1, 2, 3, 4,  
#> 5, 6, 7, 8, 9, 10, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 1, 2, 3, 4,  
#> 5, 6, 7, 8, 9, 10, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 1, 2, 3, 4,  
#> 5, 6, 7, 8, 9, 10, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 1, 2, 3, 4,  
#> 5, 6, 7, 8, 9, 10, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10)), class = "data.frame", row.names = c(NA,  
#> -200L))
#>       AIC       BIC    logLik  deviance  df.resid 
#>  771.4080  787.8996 -380.7040  761.4080       195 
#> Random effects:
#>  Groups Name        Std.Dev.  Corr
#>  group  (Intercept) 0.000e+00     
#>         x           6.065e-12  NaN
#> Number of obs: 200, groups:  group, 10
#> Fixed Effects:
#> (Intercept)            x  
#>      0.9677       1.0416  
#> optimizer (bobyqa) convergence code: 0 (OK) ; 0 optimizer warnings; 1 lme4 warnings