Linear Mixed Effect Model

Compute a linear mixed effect model (i.e., hierarchical linear model, multilevel linear model) using the nlme::lme (Pinheiro, 2006) or the lmerTest::lmer (Kuznetsova, 2017) function. The model support all dplyr::select syntax (see ?dplyr::select or vignette('mixed_effect_model) for detailed description). It will also automatically remove the response variable and id from other variables (see details below for an example). It currently only support two-level model. More complicated model structure is supported by passing the model argument.

lme_model(
  data,
  model = NULL,
  response_variable,
  level_1_factors,
  level_2_factors = NULL,
  two_way_interaction_factor = NULL,
  three_way_interaction_factor = NULL,
  id,
  estimation_method = "REML",
  opt_control = "optim",
  na.action = stats::na.exclude,
  use_package = "nlme",
  quite = F,
  ...
)

Arguments

data
data frame

model
lme4 model syntax.

response_variable
character or vector of length 1

level_1_factors
vector. Level-1 variables (e.g., individual-level)

level_2_factors
optional vector. level-2 variables (e.g., group-level)

two_way_interaction_factor
optional vector of length more than 2. Default to null

three_way_interaction_factor
optional vector of length 3. Do not include two-way interaction factors if this is not null. Default to null

id
character or vector of length 1. The nesting variable (e.g. group)

estimation_method
character. ML or REML default to REML.

opt_control
character. default to optim for lme and bobyqa for lmerTest

na.action
default to stats::na.exclude.

use_package
character. nlme or lmerTest, or lme4. Default is nlme.

quite
default to F. If set to T, it will not print the fitting model statement

...
Internal use only. It doesn't work in other cases

Value

An object of class lme of lmerModLmerTest representing the linear mixed-effects model fit.

References

Kuznetsova, A., Brockhoff, P. B., & Christensen, R. H. (2017). lmerTest package: tests in linear mixed effects models. Journal of statistical software, 82(13), 1-26.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel

Pinheiro, J., Bates, D., DebRoy, S., Sarkar, D., & Team, R. C. (2006). nlme: Linear and nonlinear mixed effects models. R package version, 3(4), 109.

Examples

# two-level model with level-1 and level-2 variable with random intercept and random slope
fit1 = lme_model(
  response_variable = JS_Individual,
  level_1_factors = Age_Individual,
  level_2_factors = contains("Country"),
  id = Country,
  data = EWCS_2015_shorten,
)

#> Fitting Model with lme: 
#>  Fixed = JS_Individual ~ Age_Individual + Hofstede_IC_Country + Hofstede_UA_Country + Hofstede_PD_Country + Hofstede_MF_Country + Hofstede_LO_Country + Hofstede_IR_Country 
#>  Random = ~ 1 + Age_Individual | Country 

# added two-way interaction factor
fit2 = lme_model(
  response_variable = JS_Individual,
  level_1_factors = Age_Individual,
  level_2_factors = contains("Country"),
  two_way_interaction_factor = c("Age_Individual", "Hofstede_IC_Country"),
  id = Country,
  data = EWCS_2015_shorten
)

#> Fitting Model with lme: 
#>  Fixed = JS_Individual ~ Age_Individual + Hofstede_IC_Country + Hofstede_UA_Country + Hofstede_PD_Country + Hofstede_MF_Country + Hofstede_LO_Country + Hofstede_IR_Country + Age_Individual*Hofstede_IC_Country 
#>  Random = ~ 1 + Age_Individual | Country 

# \donttest{
# fit3 is equivalent to fit4
fit3 = lme_model(
  response_variable = JS_Individual,
  level_1_factors = c(contains("Individual"), -JS_Individual), # !look at the difference here
  level_2_factors = contains("Country"),
  id = Country,
  data = EWCS_2015_shorten
)

#> Fitting Model with lme: 
#>  Fixed = JS_Individual ~ Gender_Individual + Age_Individual + Education_Individual + SWB_Individual + Hofstede_IC_Country + Hofstede_UA_Country + Hofstede_PD_Country + Hofstede_MF_Country + Hofstede_LO_Country + Hofstede_IR_Country 
#>  Random = ~ 1 + Gender_Individual + Age_Individual + Education_Individual + SWB_Individual | Country 

# }
# Equivalent to the above example.
# It shows that you don't need to remove response_variable from tidyselect syntax.
# \donttest{
fit4 = lme_model(
  response_variable = JS_Individual,
  level_1_factors = contains("Individual"), # !look at the difference here
  level_2_factors = contains("Country"),
  id = Country,
  data = EWCS_2015_shorten
)

#> Fitting Model with lme: 
#>  Fixed = JS_Individual ~ Gender_Individual + Age_Individual + Education_Individual + SWB_Individual + Hofstede_IC_Country + Hofstede_UA_Country + Hofstede_PD_Country + Hofstede_MF_Country + Hofstede_LO_Country + Hofstede_IR_Country 
#>  Random = ~ 1 + Gender_Individual + Age_Individual + Education_Individual + SWB_Individual | Country 

# }