Measurement Invariance

Compute the measurement invariance model (i.e., measurement equivalence model) using multi-group confirmatory factor analysis (MGCFA; Jöreskog, 1971). This function uses the lavaan::cfa (Rosseel, 2012). Users can run the configural-metric or the configural-metric-scalar comparisons (see below for detail instruction).

measurement_invariance(
  data,
  ...,
  model = NULL,
  group,
  ordered = F,
  group_partial = NULL,
  invariance_level = "scalar",
  digits = 3
)

Arguments

data
data frame

...
CFA items. Multi-factor CFA items should be separated by comma (as different argument). See below for example

model
explicit lavaan model. Either the model argument or the items argument must be specified.

group
character. group variable.

ordered
logical. default is F. If it is set to T, lavaan will treat it as a ordinal variable and use DWLS instead of ML

group_partial
items for partial equivalence. The form should be c('DV =~ item1', 'DV =~ item2').

invariance_level
"metric" or "scalar". Default is 'metric'. Set as 'metric' for configural-metric comparison, and set as 'scalar' for configural-metric-scalar comparison.

digits
number of digit to round

Value

return the compareFit object

Details

All argument must be explicitly specified. If not, all arguments will be treated as CFA items.

References

Jöreskog, K. G. (1971). Simultaneous factor analysis in several populations. Psychometrika, 36(4), 409-426.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel

Rosseel Y (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1–36. https://www.jstatsoft.org/v48/i02/.

Examples

# REMEMBER, YOU MUST NAMED ALL ARGUMENT EXCEPT THE CFA ITEMS ARGUMENT
# Fitting a multiple-factor measurement invariance model by passing items. 
measurement_invariance(
  x1:x3,
  x4:x6,
  x7:x9,
  data = lavaan::HolzingerSwineford1939,
  group = "school"
)

#> Computing CFA using:
#>   DV1 =~ x1 + x2 + x3
#>   DV2 =~ x4 + x5 + x6
#>   DV3 =~ x7 + x8 + x9
#>  [1] "Computing for configural model"
#> [1] "Computing for metric model"
#> [1] "Computing for scalar model"

#>                   chisq     df pvalue    cfi  rmsea  srmr    tli      aic
#> configural      115.851 48.000  0.000  0.923  0.097 0.068  0.885 7484.395
#> metric          124.044 54.000  0.000  0.921  0.093 0.072  0.895 7480.587
#> scalar          164.103 60.000  0.000  0.882  0.107 0.082  0.859 7508.647
#> .                                                                        
#> metric - config   8.192  6.000  0.000 -0.002 -0.004 0.004  0.009   -3.808
#> scalar - metric  40.059  6.000  0.000 -0.038  0.015 0.011 -0.036   28.059
#>                      bic     bic2
#> configural      7706.822 7516.536
#> metric          7680.771 7509.514
#> scalar          7686.588 7534.359
#> .                                
#> metric - config  -26.050   -7.022
#> scalar - metric    5.817   24.845

# Fitting measurement invariance model by passing explicit lavaan model
# I am also going to only test for metric invariance instead of the default scalar invariance 
measurement_invariance(
  model = "visual  =~ x1 + x2 + x3;
           textual =~ x4 + x5 + x6;
           speed   =~ x7 + x8 + x9",
  data = lavaan::HolzingerSwineford1939,
  group = "school",
  invariance_level = 'metric'
)

#> Computing CFA using:
#>  visual  =~ x1 + x2 + x3;
#>            textual =~ x4 + x5 + x6;
#>            speed   =~ x7 + x8 + x9[1] "Computing for configural model"
#> [1] "Computing for metric model"

#>                   chisq     df pvalue    cfi  rmsea  srmr   tli      aic
#> configural      115.851 48.000  0.000  0.923  0.097 0.068 0.885 7484.395
#> metric          124.044 54.000  0.000  0.921  0.093 0.072 0.895 7480.587
#> .                                                                       
#> metric - config   8.192  6.000  0.000 -0.002 -0.004 0.004 0.009   -3.808
#>                      bic     bic2
#> configural      7706.822 7516.536
#> metric          7680.771 7509.514
#> .                                
#> metric - config  -26.050   -7.022

if (FALSE) {
# This will fail because I did not add `model = ` in front of the lavaan model.
# Therefore,you must add the tag in front of all arguments
# For example, `return_result = 'model'` instaed of `model`
measurement_invariance(
             "visual  =~ x1 + x2 + x3;
             textual =~ x4 + x5 + x6;
             speed   =~ x7 + x8 + x9",
  data = lavaan::HolzingerSwineford1939
)
}