Model Summary for Mixed Effect Model

The function will extract the relevant coefficients from the linear mixed effect models (see supported model below).

model_summary(
  model,
  model_performance = c("R2_fixed_effect", "R2_full_model"),
  estimate_round = 3,
  p_value_round = 3,
  performance_round = 3,
  model_fit = "none"
)

Arguments

model
an object from nlme::lme, lmerTest::lmer, or lme4::glmer

model_performance
vector. R2_full_model for conditional R2 and R2_fixed_effect for marginal R2 (Nakagawa, 2013). icc for intraclass correlation coefficient. The function calls the performance package for R2 and ICC (Lüdecke et al., 2020).

estimate_round
numeric. number of digit to round to for the slope estimate

p_value_round
numeric. number of digit to round to for the p-value

performance_round
numeric. number of digit to round to for the performance statistics (e.g., R2)

model_fit
if you know your model produce same result as one of the object specify in the model argument, then you can try pass it to the function. Do it at your own risk.

Value

a data frame with estimate, degree of freedom, p_value, and whether the p-value is significant

References

Lüdecke, D., Makowski, D., Waggoner, P., Patil I (2020). performance: Assessment of Regression Models Performance. CRAN. doi: 10.5281/zenodo.3952174, R package, https://easystats.github.io/performance/.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel

Nakagawa, S., & Schielzeth, H. (2013). A general and simple method for obtaining R2 from generalized linear mixed-effects models. Methods in Ecology and Evolution, 4(2), 133–142. https://doi.org/10.1111/j.2041-210x.2012.00261.x

Examples

fit <- lme_model(
  response_variable = "Reaction",
  level_1_factors = "Days",
  id = "Subject",
  data = lme4::sleepstudy
)

#> Warning: The following columns are coerced into numeric: Subject

#> Fitting Model with lme: 
#>  Fixed = Reaction ~ Days 
#>  Random = ~ 1 + Days | Subject 

model_summary(fit)

#>          variable estimate  DF p_value significant
#> 1     (Intercept)  251.405 161   0.000         ***
#> 2            Days   10.467 161   0.000         ***
#> 3   R2_full_model    0.799                        
#> 4 R2_fixed_effect    0.279