Z-scored for multilevel analyses It will group mean center and z-scored the L1 score and create a mean score for each L2 group.

Z-scored for multilevel analyses It will group mean center and z-scored the L1 score and create a mean score for each L2 group.

z_scored_mlm_categorical(
  data,
  effect_coded,
  dummy_coded,
  group,
  keep_original = TRUE
)

Arguments

data

dataframe

effect_coded

effect coded variables for L1 group-mean centering

dummy_coded

dummy-coded variables for L2 aggregated means

group

the grouping variable. Must be character

keep_original

default is `FALSE`. Set to `TRUE` to keep original columns

Value

return a dataframe with the columns z-scored (replace existing columns)

Examples

iris %>% dplyr::mutate(cluster = rep(c(1,2,3),nrow(.)/3)) %>% effect_coding('Species',factor = FALSE) %>% dummy_coding('Species') %>% z_scored_mlm_categorical(effect_coded = dplyr::ends_with('eff'),dummy_coded = dplyr::ends_with('dum'),group = 'cluster')
#> Joining with `by = join_by(Species)`
#> Joining with `by = join_by(Species)`
#> # A tibble: 150 × 14
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species cluster
#>           <dbl>       <dbl>        <dbl>       <dbl> <chr>     <dbl>
#>  1          5.1         3.5          1.4         0.2 setosa        1
#>  2          4.9         3            1.4         0.2 setosa        2
#>  3          4.7         3.2          1.3         0.2 setosa        3
#>  4          4.6         3.1          1.5         0.2 setosa        1
#>  5          5           3.6          1.4         0.2 setosa        2
#>  6          5.4         3.9          1.7         0.4 setosa        3
#>  7          4.6         3.4          1.4         0.3 setosa        1
#>  8          5           3.4          1.5         0.2 setosa        2
#>  9          4.4         2.9          1.4         0.2 setosa        3
#> 10          4.9         3.1          1.5         0.1 setosa        1
#> # ℹ 140 more rows
#> # ℹ 8 more variables: Species_setosa_eff <dbl>, Species_versicolor_eff <dbl>,
#> #   Species_setosa_dum <dbl>, Species_versicolor_dum <dbl>,
#> #   Species_setosa_eff_group_c <dbl>, Species_versicolor_eff_group_c <dbl>,
#> #   Species_setosa_dum_mean_z <dbl>, Species_versicolor_dum_mean_z <dbl>