Confirmatory Factor Analysis

The function fits a CFA model using the lavaan::cfa (Rosseel, 2012) function. Users can fit single and multiple factors CFA (read details), and it also supports multilevel CFA (specifying the group). Users can pass the items (see below for example) or an explicit lavaan model for more versatile usage.

cfa_summary(
  data,
  ...,
  model = NULL,
  group = NULL,
  summary_item = c("cfi", "rmsea", "tli", "srmr"),
  ordered = F,
  return_result = "model",
  jmv_result = F,
  quite = F,
  group_partial = NULL
)

Arguments

data
data frame

...
CFA items. Multi-factor CFA items should be separated by comma (as different argument). See below for example

model
explicit lavaan model. Must be specify with model = lavaan_model_syntax.

group
optional character. used for multi-level CFA. the nested variable for multilevel dataset (e.g., Country)

summary_item
vector of fit indices. Default is CFI, RMSEA, TLI, and SRMR.

ordered
logical. default is F. If it is set to T, lavaan will treat it as a ordinal variable and use DWLS instead of ML

return_result
default is model. Options are 'model' (lavaan model), 'short_summary' (fit index summary only), 'long_summary' (lavaan full summary).

jmv_result
default is F. Print a good-lookng summary use jmv::cfa. It currently only support one-level CFA with latent factors and latent variable. It does not support complex lavaan structure at the momenet. I will work on supporting residual covariance.

quite
default as F. If set to true, it will not print the running model statement.

group_partial
Items for partial equivalence. The form should be c('DV =~ item1', 'DV =~ item2').

Details

All argument must be explicitly specified. If not, all arguments will be treated as CFA items.

References

Bao, H.-W.-S. (2021). bruceR: Broadly useful convenient and efficient R functions. R package version 0.6.0. https://CRAN.R-project.org/package=bruceR

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel

Rosseel Y (2012). lavaan: An R Package for Structural Equation Modeling. Journal of Statistical Software, 48(2), 1–36. https://www.jstatsoft.org/v48/i02/.

Examples

# REMEMBER, YOU MUST NAMED ALL ARGUMENT EXCEPT THE CFA ITEMS ARGUMENT
# Fitting a single factor CFA model
fit = cfa_summary(
  data = lavaan::HolzingerSwineford1939,
  x1:x3,
  jmv_result = TRUE
)

#> Computing CFA using:
#>   DV1 =~ x1 + x2 + x3
#>  
#>  CONFIRMATORY FACTOR ANALYSIS
#> 
#>  Factor Loadings                                                                                 
#>  ─────────────────────────────────────────────────────────────────────────────────────────────── 
#>    Factor    Indicator    Estimate     SE            Z           p             Stand. Estimate   
#>  ─────────────────────────────────────────────────────────────────────────────────────────────── 
#>    DV1       x1           0.7236898    0.08998041    8.042749    < .0000001          0.6209312   
#>              x2           0.5629087    0.08221199    6.847040    < .0000001          0.4788703   
#>              x3           0.8013092    0.09289239    8.626210    < .0000001          0.7096890   
#>  ─────────────────────────────────────────────────────────────────────────────────────────────── 
#> 
#> 
#>  FACTOR ESTIMATES
#> 
#>  Factor Covariances                                                      
#>  ─────────────────────────────────────────────────────────────────────── 
#>                  Estimate      SE    Z    p            Stand. Estimate   
#>  ─────────────────────────────────────────────────────────────────────── 
#>    DV1    DV1    1.000000 ᵃ                                              
#>  ─────────────────────────────────────────────────────────────────────── 
#>    ᵃ fixed parameter
#> 
#> 
#>  MODEL FIT
#> 
#>  Test for Exact Fit              
#>  ─────────────────────────────── 
#>    χ²          df    p           
#>  ─────────────────────────────── 
#>    0.000000     0                
#>  ─────────────────────────────── 
#> 
#> 
#>  Fit Measures                                                 
#>  ──────────────────────────────────────────────────────────── 
#>    CFI         TLI         RMSEA       Lower       Upper      
#>  ──────────────────────────────────────────────────────────── 
#>    1.000000    1.000000    0.000000    0.000000    0.000000   
#>  ──────────────────────────────────────────────────────────── 
#> 

# Fitting a multilevel single factor CFA model
fit = cfa_summary(
  data = lavaan::HolzingerSwineford1939,
  x1:x3,
  group = "sex"
)

#> Computing CFA using:
#>   DV1 =~ x1 + x2 + x3
#>  

# Fitting a multiple factor CFA model
fit = cfa_summary(
  data = lavaan::HolzingerSwineford1939,
  x1:x3,
  x4:x6,
  x7:x9,
)

#> Computing CFA using:
#>   DV1 =~ x1 + x2 + x3
#>   DV2 =~ x4 + x5 + x6
#>   DV3 =~ x7 + x8 + x9
#>  

# Fitting a CFA model by passing explicit lavaan model (equivalent to the above model)
# Note in the below function how I added `model = ` in front of the lavaan model.
# Similarly, the same rule apply for all arguments (e.g., `ordered = F` instead of `F`)
fit = cfa_summary(
 model = "visual  =~ x1 + x2 + x3;textual =~ x4 + x5 + x6;",
 data = lavaan::HolzingerSwineford1939,
 jmv_result = TRUE
)

#> Computing CFA using:
#>  visual  =~ x1 + x2 + x3;textual =~ x4 + x5 + x6;
#>  CONFIRMATORY FACTOR ANALYSIS
#> 
#>  Factor Loadings                                                                                   
#>  ───────────────────────────────────────────────────────────────────────────────────────────────── 
#>    Factor     Indicator    Estimate     SE            Z            p             Stand. Estimate   
#>  ───────────────────────────────────────────────────────────────────────────────────────────────── 
#>    visual     x1           0.9066191    0.09275204     9.774654    < .0000001          0.7778857   
#>               x2           0.5067589    0.08648902     5.859228    < .0000001          0.4311034   
#>               x3           0.6418333    0.08131033     7.893625    < .0000001          0.5684472   
#>    textual    x4           0.9905773    0.05670103    17.470183    < .0000001          0.8523434   
#>               x5           1.1005015    0.06262075    17.574071    < .0000001          0.8542099   
#>               x6           0.9166619    0.05377935    17.044867    < .0000001          0.8380666   
#>  ───────────────────────────────────────────────────────────────────────────────────────────────── 
#> 
#> 
#>  FACTOR ESTIMATES
#> 
#>  Factor Covariances                                                                               
#>  ──────────────────────────────────────────────────────────────────────────────────────────────── 
#>                          Estimate       SE            Z           p             Stand. Estimate   
#>  ──────────────────────────────────────────────────────────────────────────────────────────────── 
#>    visual     visual     1.0000000 ᵃ                                                              
#>               textual    0.4606289      0.06300850    7.310583    < .0000001          0.4606289   
#>    textual    textual    1.0000000 ᵃ                                                              
#>  ──────────────────────────────────────────────────────────────────────────────────────────────── 
#>    ᵃ fixed parameter
#> 
#> 
#>  MODEL FIT
#> 
#>  Test for Exact Fit              
#>  ─────────────────────────────── 
#>    χ²          df    p           
#>  ─────────────────────────────── 
#>    24.36130     8    0.0019929   
#>  ─────────────────────────────── 
#> 
#> 
#>  Fit Measures                                                        
#>  ─────────────────────────────────────────────────────────────────── 
#>    CFI          TLI          RMSEA         Lower         Upper       
#>  ─────────────────────────────────────────────────────────────────── 
#>    0.9749691    0.9530670    0.08242913    0.04626573    0.1208571   
#>  ─────────────────────────────────────────────────────────────────── 
#> 

if (FALSE) {
# This will fail because I did not add `model = ` in front of the lavaan model.
# Therefore,you must add the tag in front of all arguments
# For example, `return_result = 'model'` instaed of `model`
cfa_summary("visual  =~ x1 + x2 + x3
             textual =~ x4 + x5 + x6
             speed   =~ x7 + x8 + x9 ",
  data = lavaan::HolzingerSwineford1939
)
}