Generalized Linear Mixed Effect Model

Generalized linear mixed effect model. The function uses the lme4::glmer(Bates et al., 2014) function. This function is still in very early development stage.

glme_model(
  data,
  response_variable,
  level_1_factors,
  level_2_factors = NULL,
  family,
  two_way_interaction_factor = NULL,
  three_way_interaction_factor = NULL,
  id,
  estimation_method = "REML",
  opt_control = "bobyqa",
  na.action = stats::na.exclude,
  quite = F
)

Arguments

data
data frame

response_variable
character or vector of length 1

level_1_factors
vector. Level-1 variables (e.g., individual-level)

level_2_factors
optional vector. level-2 variables (e.g., group-level)

family
a GLM family. It will passed to the family argument in glmer. See ?glmer for possible options.

two_way_interaction_factor
optional vector of length more than 2. Default to null

three_way_interaction_factor
optional vector of length 3. Do not include two-way interaction factors if this is not null. Default to null

id
character or vector of length 1. The nesting variable (e.g. group)

estimation_method
character. ML or REML default to REML.

opt_control
character. default to bobyqa

na.action
default to stats::na.exclude.

quite
default to F. If set to T, it will not print the fitting model statement

Value

An object of class "glmerMod" representing the linear mixed-effects model fit.

References

Bates, D., Mächler, M., Bolker, B., & Walker, S. (2014). Fitting Linear Mixed-Effects Models Using lme4. Journal of Statistical Software, 67(1), 1–48. doi: 10.18637/jss.v067.i01.

Moy, J. H. (2021). psycModel: Integrated Toolkit for Psychological Analysis and Modeling in R. R package. https://github.com/jasonmoy28/psycModel

Examples

x <- rnorm(200)
y <- rpois(200, exp(1 + x))
group <- rep(1:10, 20)
test_df <- as.data.frame(cbind(x, y, group))
fit = glme_model(
  data = test_df,
  response_variable = "y",
  level_1_factors = "x",
  family = poisson(link = "log"),
  id = "group"
)

#> Fitting Model with glmer: 
#>  Formula = y ~ x + (1 + x | group)
#>  Family = poisson

#> boundary (singular) fit: see ?isSingular