Convert HTML to PDF

This is a helper function for knitting Rmd. Due to technological limitation, the output cannot knit to PDF in Rmd directly. It uses the pagedown::chrome_print() in the backend. You must first knit to HTML, then you can use this function to covert them to PDF if you wish. I am sorry. I can't find any way to implement it. If you know how to implement it, please submit an issue on github to let me know.

html_to_pdf(file_path = NULL, dir = NULL, render_exist = F)

Arguments

file_path
file path to the HTML file (can be relative if you are in a R project)

dir
file path to the directory of all HTML files (can be relative if you are in a R project)

render_exist
overwrite exist PDF. Default is F

Value

no return

Examples

if (FALSE) {
html_to_pdf(file_path = "html_name.html")
# all HTML files in the my_html_folder will be converted
html_to_pdf(dir = "Users/Desktop/my_html_folder")
}