Exploratory Linear Regression Model Table

Exploratory analyses for linear regression models with multiple response, predictor, and two-way interaction variables. (lm models). At the moment, multi-categorical variables are not supported as predictors or interactions (but control is fine). Binary variable should be numeric instead of factor

lm_model_explore(
  data,
  response_variable,
  predictor_variable,
  two_way_interaction_variable = NULL,
  control_variable = NULL,
  marginal_alpha = 0.1,
  return_result = FALSE,
  print_control = FALSE,
  verbose = TRUE,
  show_p = TRUE
)

Arguments

data

data.frame

response_variable

Response variable. Support dplyr::select() syntax.

predictor_variable

Pred variable. Support dplyr::select() syntax.

two_way_interaction_variable

Two-way interaction variable. Each two-way interaction variable will interact with each pred variable. Support dplyr::select() syntax.

control_variable

Control variables. Support dplyr::select() syntax.

marginal_alpha

Set marginal_alpha level for marginally significant (denoted by .). Set to 0.05 if do not want marginally significant denotation.

return_result

Default is FALSE. If TRUE, it returns the model estimates as a data frame.

print_control

Default is FALSE. If TRUE, print coefficients of control variables.

verbose

Default is TRUE. Set to FALSE to suppress outputs

show_p

Default is TRUE. When TRUE, show the p-value in parenthesis.

Value

data.frame

Examples

test = data.frame(y1 = rnorm(1000,2,3),
y2 = rnorm(1000,10,2),
y3 = rnorm(1000,1,4),
x1 = rnorm(1000,100,10),
x2 = rnorm(1000,10,1),
x3 = rnorm(1000,6,2),
m1 = rnorm(1000,3,1),
m2 = rnorm(1000,2,0.5),
m3 = rnorm(1000,9,0.1),
c1 = rnorm(1000,5,0.4),
c2 = rnorm(1000,2,0.2),
c3 = rnorm(1000,7,0.9)
)

lm_model_explore(data = test, 
                 response_variable = c(y1,y2,y3),
                 predictor_variable = c(x1,x2,x3),
                 two_way_interaction_variable = c(m1,m2,m3),
                 control_variable = c(c1,c2,c3))
#> ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
#>   Response  Pred  Interact_pred  Interact_term   Interact_term_coef   Interact_pred_coef            Pred_coef          (Intercept)       df     r2
#> ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
#>         y1    x1             m1          x1*m1    0.004 (0.650)       -0.519 (0.574)        0.004 (0.890)        0.200 (0.951)      993.000  0.006
#>         y1    x1             m2          x1*m2    0.026 (0.161)       -2.799 (0.141)       -0.038 (0.340)        4.380 (0.307)      993.000  0.007
#>         y1    x1             m3          x1*m3   -0.068 (0.479)        6.843 (0.483)        0.625 (0.468)      -62.799 (0.474)      993.000  0.005
#>         y1    x2             m1          x2*m1    0.125 (0.223)       -1.355 (0.189)       -0.300 (0.341)        3.702 (0.291)      993.000  0.005
#>         y1    x2             m2          x2*m2    0.173 (0.355)       -1.871 (0.321)       -0.277 (0.467)        3.488 (0.397)      993.000  0.004
#>         y1    x2             m3          x2*m3    1.247 (0.191)      -12.339 (0.194)      -11.155 (0.193)      111.023 (0.195)      993.000  0.004
#>         y1    x3             m1          x3*m1    0.024 (0.622)       -0.250 (0.431)       -0.051 (0.733)        0.994 (0.595)      993.000  0.003
#>         y1    x3             m2          x3*m2    0.132 (0.156)       -0.922 (0.115)       -0.238 (0.205)        1.976 (0.311)      993.000  0.004
#>         y1    x3             m3          x3*m3    0.298 (0.493)       -1.783 (0.518)       -2.654 (0.497)       16.338 (0.510)      993.000  0.002
#>         y2    x1             m1          x1*m1    0.008 (0.210)       -0.762 (0.226)       -0.029 (0.126)       11.857 (0.000) ***  993.000  0.005
#>         y2    x1             m2          x1*m2   -0.003 (0.824)        0.206 (0.874)       -0.001 (0.980)        9.216 (0.002) **   993.000  0.004
#>         y2    x1             m3          x1*m3   -0.001 (0.993)        0.034 (0.996)       -0.001 (0.998)        9.355 (0.876)      993.000  0.004
#>         y2    x2             m1          x2*m1    0.064 (0.356)       -0.615 (0.379)       -0.422 (0.048) *     12.994 (0.000) ***  993.000  0.017
#>         y2    x2             m2          x2*m2    0.093 (0.464)       -1.023 (0.423)       -0.417 (0.106)       13.186 (0.000) ***  993.000  0.017
#>         y2    x2             m3          x2*m3   -0.600 (0.352)        5.781 (0.369)        5.170 (0.373)      -40.906 (0.480)      993.000  0.017
#>         y2    x3             m1          x3*m1   -0.003 (0.919)        0.045 (0.835)        0.037 (0.715)        8.628 (0.000) ***  993.000  0.003
#>         y2    x3             m2          x3*m2   -0.030 (0.639)        0.090 (0.821)        0.085 (0.504)        8.589 (0.000) ***  993.000  0.004
#>         y2    x3             m3          x3*m3   -0.175 (0.554)        1.006 (0.592)        1.602 (0.547)       -0.296 (0.986)      993.000  0.004
#>         y3    x1             m1          x1*m1    0.016 (0.208)       -1.608 (0.206)       -0.026 (0.509)        1.207 (0.787)      993.000  0.010
#>         y3    x1             m2          x1*m2   -0.036 (0.165)        3.752 (0.152)        0.094 (0.085) .    -10.947 (0.064) .    993.000  0.010
#>         y3    x1             m3          x1*m3   -0.059 (0.652)        6.349 (0.636)        0.554 (0.640)      -60.464 (0.617)      993.000  0.008
#>         y3    x2             m1          x2*m1   -0.083 (0.555)        0.817 (0.565)        0.357 (0.412)       -4.643 (0.337)      993.000  0.007
#>         y3    x2             m2          x2*m2   -0.084 (0.744)        0.991 (0.703)        0.278 (0.596)       -4.189 (0.460)      993.000  0.007
#>         y3    x2             m3          x2*m3    0.588 (0.654)       -5.407 (0.680)       -5.178 (0.661)       46.479 (0.694)      993.000  0.007
#>         y3    x3             m1          x3*m1   -0.029 (0.674)        0.161 (0.712)        0.142 (0.492)       -2.096 (0.416)      993.000  0.007
#>         y3    x3             m2          x3*m2    0.265 (0.038) *     -1.437 (0.074) .     -0.460 (0.075) .      0.928 (0.729)      993.000  0.011
#>         y3    x3             m3          x3*m3   -0.070 (0.907)        0.766 (0.840)        0.686 (0.899)       -8.484 (0.804)      993.000  0.007
#> ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
#> Note: Coefficient (p-value): . p < 0.1, * p < 0.05, ** p < 0.01, *** p < 0.001